refactored to use TaskService
